email-order-agent/requirements_vm.txt
# Core dependencies for Email Order Agent
anthropic==0.69.0
twilio>=9.0.0
python-dotenv==1.0.0
pdfminer.six==20231228
pdfplumber==0.11.0
schedule==1.2.2
requests==2.32.5
cryptography==46.0.2

# Optional but recommended
Pillow==11.3.0
